archivers/xmill: fix build with clang
PR:		191372
Submitted by:	tkato432 yahoo com

+++ archivers/xmill/files/patch-XMill-CompressMan.cpp
@@ -5,7 +5,7 @@
     while(compressor!=NULL)
     {
 -      printf("%lu =>",(unsigned int)compressor);
-+      printf("%lu =>",(unsigned intptr_t)compressor);
++      printf("%lu =>",(unsigned)(intptr_t)compressor);
        printf("%s\n",compressor->GetName());
        compressor=compressor->next;
     }
